requirements.txt
ta
pandas
pytz
python-binance
ccxt
